+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+* **Deprecation** The method `read_u32_from_heap` has
+
 ## [0.5.8](https://github.com/streamingfast/substreams-rs/release/tag/v0.5.8)
 
 ### Highlights